:recycle: Move impersonation claims to their own provider
