+++ frontend/src/App/Layout/Header/index.tsx
@@ -8,14 +8,14 @@ import { SessionMethod, useSessionDispatch } from "src/hooks/useSession";
 import { HasuraUserRole } from "src/types";
 import { GetPaymentRequestIdsQuery } from "src/__generated/graphql";
 import View from "./View";
-import { useTokenSet } from "src/hooks/useTokenSet";
+import { useImpersonationClaims } from "src/hooks/useImpersonationClaims";
 
 export default function Header() {
   const location = useLocation();
   const { isLoggedIn, githubUserId } = useAuth();
   const { T } = useIntl();
   const dispatchSession = useSessionDispatch();
-  const { impersonationSet } = useTokenSet();
+  const { impersonationSet } = useImpersonationClaims();
   const impersonating = !!impersonationSet;
 
   const { data: paymentRequestIdsQueryData } = useHasuraQuery<GetPaymentRequestIdsQuery>(
